+++ src/main/java/kitchenpos/menus/application/MenuService.java
@@ -1,8 +1,8 @@
 package kitchenpos.menus.application;
 
-import kitchenpos.menus.domain.Menu;
 import kitchenpos.menus.domain.MenuGroupRepository;
 import kitchenpos.menus.domain.MenuRepository;
+import kitchenpos.menus.tobe.domain.Menu;
 import kitchenpos.menus.tobe.domain.MenuGroup;
 import kitchenpos.menus.tobe.domain.MenuProduct;
 import kitchenpos.products.domain.Product;
