+++ src/main/java/kitchenpos/products/application/ProductService.java
@@ -1,7 +1,7 @@
 package kitchenpos.products.application;
 
-import kitchenpos.menus.domain.Menu;
 import kitchenpos.menus.domain.MenuRepository;
+import kitchenpos.menus.tobe.domain.Menu;
 import kitchenpos.menus.tobe.domain.MenuProduct;
 import kitchenpos.products.domain.Product;
 import kitchenpos.products.domain.ProductRepository;
