+++ src/test/java/kitchenpos/menus/application/InMemoryMenuRepository.java
@@ -1,9 +1,14 @@
 package kitchenpos.menus.application;
 
-import kitchenpos.menus.domain.Menu;
 import kitchenpos.menus.domain.MenuRepository;
+import kitchenpos.menus.tobe.domain.Menu;
 
-import java.util.*;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
+import java.util.UUID;
 import java.util.stream.Collectors;
 
 public class InMemoryMenuRepository implements MenuRepository {
